Fix compiler warning: implicit conversion loses integer precision
src/utils_dates.c:518:20: warning: implicit conversion loses integer precision: 'gsize' (aka 'unsigned long') to 'guint' (aka 'unsigned int') [-Wshorten-64-to-32]
        longueur = g_date_strftime ( retour_str, SIZEOF_FORMATTED_STRING_DATE, format, date );
                 ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
src/utils_dates.c:546:16: warning: implicit conversion loses integer precision: 'gsize' (aka 'unsigned long') to 'guint' (aka 'unsigned int') [-Wshorten-64-to-32]
    longueur = g_date_strftime ( retour_str, SIZEOF_FORMATTED_STRING_DATE, "%m/%d/%Y", date );
             ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/utils_dates.c
@@ -509,7 +509,7 @@ gchar *gsb_format_date ( gint day, gint month, gint year )
 gchar *gsb_format_gdate ( const GDate *date )
 {
     gchar retour_str[SIZEOF_FORMATTED_STRING_DATE];
-    guint longueur = 0;
+    gsize longueur = 0;
 
     if ( !date || !g_date_valid ( date ) )
         return my_strdup ( "" );
@@ -536,7 +536,7 @@ gchar *gsb_format_gdate ( const GDate *date )
 gchar *gsb_format_gdate_safe ( const GDate *date )
 {
     gchar retour_str[SIZEOF_FORMATTED_STRING_DATE];
-    guint longueur;
+    gsize longueur;
 
     if ( !date || !g_date_valid ( date ) )
     {
